Geometry test refactor.
I refactored the geometry tests some. Split test_geometry.py. Some pep8 refactoring. Minimized geometry test imports, minimized some of the global variables in geometry tests.
